// src/contexts/auth/oauth/types.ts
/*
 * Copyright (c) 2025 FLECS Technologies GmbH
 *
 * Created on Nov 10 2025
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import * as oauth from 'oauth4webapi';

export interface RealmAccess {
  roles: string[];
}

export interface ResourceAccess {
  [resource: string]: {
    roles: string[];
  };
}

export interface User {
  sub: string;
  email?: string;
  name?: string;
  preferred_username?: string;
  exp?: number;
  realm_access?: RealmAccess;
  resource_access?: ResourceAccess;
  access_token?: string;
  [key: string]: any; // Allow additional properties
}

export interface JWTHeader {
  alg: string;
  kid?: string;
  typ?: string;
}

export interface AuthState {
  isAuthenticated: boolean;
  isLoading: boolean;
  user: User | null;
  error: Error | null;
}

export interface AuthContextValue extends AuthState {
  signIn: () => Promise<void>;
  signOut: () => Promise<void>;
  clearError: () => void;
  handleOAuthCallback: () => Promise<void>;
  refreshAuthState: () => Promise<void>;
  isConfigReady: boolean;
}

export interface OAuthConfig {
  issuer: string;
  redirect_uri: string;
  scope: string;
  kind: 'oauth' | 'oidc';
  client_secret?: string;
  post_logout_redirect_uri?: string;
}

export interface OAuthState {
  config: OAuthConfig | null;
  authServer: oauth.AuthorizationServer | null;
  client: oauth.Client | null;
}